+++ CHANGELOG.md
@@ -27,6 +27,11 @@
 
 ### Added
 
+- [#2741](https://github.com/ChainSafe/forest/issues/2741): Support importing
+  zstd compressed snapshot car files
+- [#2741](https://github.com/ChainSafe/forest/issues/2741): Support fetching
+  zstd compressed snapshots with filecoin provider via `--compressed` option
+
 ### Changed
 
 ### Removed